W203: late first divergence — +2 MATCH (6530 -> 6532, 79.6%), nd -1474, one window cleared
Census build/wip/W203_localised.md ranks the 281 first-party NONMATCHING functions whose
FIRST differing word sits at byte offset 64 or later, descending.  A large `first@` means
the prologue, the frame and everything before it are already byte-identical and the whole
defect lives after that point — the cheapest localisation available.  0 MATCHes lost,
0 new over-window, 0 compile errors.

New MATCHes:
  mt_evtMenu 0037baf0   nd225 -> 0.  Cases 0-2 recomputed the real scaled y after a call
                        instead of retaining it across each PAIR of draws, and the final
                        0x30-or-1 body was four instructions short because an if/OR chain
                        emits an inverted test where retail wants an ascending NESTED
                        switch with two positive tests and an out-of-line body.
  bpe_main 002496e0     nd84 -> 0.  Four node stores were staged through a temporary:
                        we resolved the pointer chain before materialising each constant
                        where retail materialises constants first.  The last twelve came
                        from the origin aggregate's store order — retail writes z=100
                        before x=y=0.

The largest reductions, all found by decoding retail from the divergence point:

  mt_evtMenu 003747a0   nd450 -> nd113.  The first case's 0x7cce50 text object was
                        materialised as an absolute lui/ori pair; retail uses the
                        GP-RELATIVE object.  That removed one instruction and pushed the
                        first divergence from offset 96 to 372.  **Note the direction** —
                        wave 202 found the opposite defect twice (GP-relative where retail
                        is absolute).  Neither form is "correct"; it is per-site.
  btlFormation 002c3430 nd249 -> nd11.  Retail checks the unit active flag BEFORE loading
                        unitData, masks candidateCount before indexing the stack candidate
                        array, and uses a fresh caller-saved counter for the final
                        call-free copy loop.
  fclCombineEffect 0041a730  nd243 -> nd51, 464/464 -> 452/464.  A u8 return contract was
                        adding retail-absent truncation, and two variable shifts carried
                        an explicit `& 0x1f` that retail does not emit.
  btlCamera 002ae740    2116/2112 OVER-WINDOW nd984 -> 2108/2112 nd849.  Retail uses the
                        high scratch pair buf[64]/buf[65], not buf[24]/buf[25], and keeps
                        the branch-local scale as an FP scalar rather than an invented
                        persistent buf[33] result; the later 0x106 access is unsigned.
  h_camp 0011cc80       nd485 -> nd30.  Retail initialises the first selected-entry packed
                        coordinate with `startX = -19.0f`; we retained the loop's previous
                        value.
  y_runtime 0044a420    nd197 -> nd124, deficit 12 -> 4.  Retail signed-loads the s16
                        fourth argument and branches directly on the comparison; we had a
                        u16 helper contract and a persisted boolean.
  mt_evtMenu 0036f000   nd142 -> nd126.  Three real call arguments were missing: the 0x924
                        object to FUN_00195020, the 0x6c child to FUN_00361890, and 7/0x9c
                        to FUN_00386b70; plus three low-memory literals needing typed
                        absolute objects.
  bp_root 00202010      nd316 -> nd238.  Retail places the zero-selected reset INLINE and
                        branches positively to an out-of-line selected body; inverting the
                        source if/else body order reproduces that layout.
  bp_tex 00221b60       nd359 -> nd336.  Five `mask = 1/2` assignments should be `mask |=`
                        — retail preserves accumulated bits with `ori`.
  mdlFile 0034c1b0      nd47 -> nd2, size now exact.  Retail executes two genuine VU0
                        `sqc2 vf0` stores at offsets 88 and 92 to re-zero the first two
                        16-byte slots after the memset.  Annotated hardware asm.
  datCalc 00306bc0      nd195 -> nd178.  The computed result was stored through param_7 by
                        retail before clearing the two unit fields; we omitted the store.
  _h_camp_data 0016c010 nd240 -> nd231 (absolute callback alias), mt_evtMain 00368f20
                        nd298 -> nd296 (four sentinel stores used unsigned 0xff/0xffff
                        where retail materialises signed -1), mt_evtEffect 00395570
                        nd58 -> nd54, btlFormation 002bf3f0 nd289 -> nd286,
                        mdlFile 003268c0 nd64 -> nd62 (u64 return contract emitting a
                        trailing dsll32/dsrl32 pair), y_runtime 00454a70 nd400 -> nd385
                        (the incoming RwV3d is a by-value ABI aggregate, not a pointer;
                        the caller 00454f50 was migrated with it).

SIX RESIDUALS PERMANENTLY CLOSED with their exact remaining words written out, which
converts a recurring cost into a settled fact:
  h_fade 00107b20 nd41 — one commutative `mul.s` orientation then a cyclic a1/a2 cascade
    from `mflo`.  k_encount 001d8e60 nd1 — a single `addu v0,v0,a0` vs `addu v0,a0,v0`;
    its genuine locals already sit at retail's slots.  btlCamera 002b17a0 nd3 — `mul.s`
    orientation at 0x1fc plus `addiu`/`daddiu` at 0x2cc.  bp_dialog 0025c220 nd2 — one
    commutative `mul.s` at 1072.  sfl_card 00258b40 nd13 — call-argument setup order.
    bsa_main 0020fe30 — seven sites, all `addu v0,v0,v1` vs `addu v0,v1,v0`.
  **_h_camp_148 00148880 settles a recurring misreading**: its retail function ENDS at
    +896 and the +900/+904/+908 tail is zero padding, so the "+12 deficit" was never three
    missing instructions.  Check for padding before treating a small deficit as logic.

Measured and reverted: datCalendarPrelude 00179360 (making empty case 32 a `return`
filled the 8-byte deficit exactly but displaced the long linear dispatch, nd274 -> 1065),
bcm_panel 002257f0, bpp_panel 00232ea0, mdlFile 003520a0 / 00349090 / 00349450 / 0033c0a0
/ 0031b680, k_sceneDraw 0019ef80, op_wait 0026e000, h_fade 001081e0, fclCombineMisc
003d64a0, fclShopMisc 00409c80, fclCombine 003c8400, datCalendar 0017fe10, datCalc
00300100, _h_camp_status 00128720 / 00128c40, mt_evtMenu 003722d0 / 00384d30 / 0038e660 /
0036fda0, y_runtime 0044a790 / 0045b8f0 / 00188c30 / 00189df0, btlCamera 002a79f0,
btlFormation 002bf690.

+++ src/Event/Fcl/fclCombineMisc.c
@@ -1051,7 +1051,6 @@ u32 FUN_003d64a0(void *param_1,u16 param_2,u16 param_3)
       memset(uStack_20,0,0x18);
 
       uVar6 = ((int)((u32)*(u8 *)(DAT_007ce420 + 3 + (u32)param_2 * 0xe) +
-
                     (u32)*(u8 *)(DAT_007ce420 + 3 + iVar7)) >> 1) + 1U & 0xff;
 
       uStack_20[0] = param_2;
@@ -1059,7 +1058,6 @@ u32 FUN_003d64a0(void *param_1,u16 param_2,u16 param_3)
       uStack_20[1] = param_3;
 
       if ((cVar3 == *(char *)(DAT_007ce420 + 2 + (u32)param_2 * 0xe)) &&
-
          (cVar3 == *(char *)(DAT_007ce420 + 2 + iVar7))) {
 
         sVar4 = FUN_003d7da0(cVar3,uVar6,uStack_20);
@@ -1450,7 +1448,7 @@ s32 FUN_003d6e60(s32 param_1,s32 param_2)
   s32 temp_2;
   s32 var_16;
   s32 var_5;
-  u16 temp_19;
+  u32 temp_19;
   u16 temp_3;
   u8 *temp_4;
   void *dest;
